CLA0003-564 - Ohai Lock Removal
What was originally an attempt to remove the oahi lock required for Clarus actually turned into a nginx and rbenv cookbook change too. The previous implementations of these cookbooks had not been updated in many years and were hitting problems with Chef 13, so I've replaced them with maintained versions. Hopefully, nothing should really change.

Upped the version of 0.5.0 to signify the change.