CLA0003-970 - Data Import On Demand
Took advantage of the upload hook that Pure FTP give us, and calling a ruby script that adds a single processing task line to resque, with a 30 seconds delay. From this, we will then be able to receive information about individual failures, and see them process in order, potentially by multiple workers.

The delay is just so EFS can catchup throughout the workers, so the file exists when try to process it, but also feels nicer to the user that the files don't immediately disappear.

You should be able to see files ready to process in the Delayed tab on Resque.

#CLA0003-970 Fixed